--- app_options.rb
@@ -1,140 +1,0 @@
-# require_relative 'student'
-# require_relative 'teacher'
-
-module AppOptions
-  def list_all_books
-    puts "No list to display" unless @books.length > 0
-    @books.each { |book|
-    puts "Title: #{book.title}, Author: #{book.author}"
-    }
-    puts "\n"
-    run
-  end
-
-  def list_people
-    puts "No list to display" unless @people.length > 0
-
-    @people.each {|person|
-    puts "[#{person.class.name}] ID: #{person.id}, Name: #{person.name}, Age: #{person.age}"
-    }
-
-    puts "\n"
-    run
-  end
-
-  def create_person
-    puts "press 1 to create a student"
-    puts "press 2 to create a teacher"
-
-    number = gets.chomp
-
-    case number
-    when "1"
-      puts "Enter student's name"
-      name = gets.chomp
-      
-      puts "Enter student's Age"
-      age = gets.chomp   
-
-      puts "Parent's Permission? [Y/N]"
-      parent_permission = gets.chomp.upcase
-      
-      
-      if parent_permission == 'N'       
-        student = Student.new('unknown', age, name, parent_permission: false)
-        @people << student
-        puts "student is under age and would require parent's permission"
-        puts "\n"
-      else 
-      
-        student = Student.new('unknown', age, name, parent_permission: true)
-        @people << student
-        puts "Student added"
-        puts "\n"
-      end
-
-    when '2'
-      puts "Enter Teacher's name"
-      name = gets.chomp
-
-      puts "Enter Teacher's Age"
-      age = gets.chomp
-      
-      puts "Enter teacher's specialization"
-      specialization = gets.chomp
-
-      teacher = Teacher.new(specialization, age, name, parent_permission: true)
-      @people << teacher
-      puts "Teacher added"
-      puts "\n"
-
-    else
-      puts "You entered wrong number"
-      puts "\n"
-      run
-    end
-  
-    run
-  end
-
-def create_book
-  puts "Enter Book title"
-  title = gets.chomp
-
-  puts "Enter Author's name"
-  author = gets.chomp
-
-  new_book = Book.new(title, author)
-  @books << new_book
-  puts "New book created"
-
-  puts "\n"
-  run
-end
-
-def create_rental
-  puts "Pick a book from the following list(Enter the serial number)"
-  @books.each_with_index { |book, index|
-  puts "#{index}) Title: #{book.title}, Author: #{book.author}"
-  }
-  book_selected = gets.chomp.to_i
-
-  puts "Select a person from the following list(Enter the serial number and not the ID)"
-  @people.each_with_index { |person, index|
-  puts "#{index})  [#{person.class.name}] Name: #{person.name}, Age: #{person.age}, ID: #{person.id}"
-  }
-
-  person_selected = gets.chomp.to_i
-
-  puts "Enter date"
-  date = gets.to_s
-
-  new_rental = Rental.new(date, @books[book_selected], @people[person_selected])
-  @rental << new_rental
-  print @rental
-  puts "Book rented successfully"
-  puts "\n"
-
-  run
-end
-
-def list_rental_per_person_id
-  puts "Enter the ID of the person"
-  person_id = gets.chomp
-
- @rental.each { |rent|
-  # puts rent.person.id
-  # puts person_id
- if rent.person.id == person_id.to_i
-  puts "Date: #{rent.date} #{rent.book.title} by #{rent.book.author}"
-  puts "\n"
- else
-  puts "Record not found"
-  puts "\n"
- end
- }
-
- run
-end
-  
-end
